requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Session==0.8.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
msgspec==0.19.0
multidict==6.3.2
mysql-connector-python==9.2.0
nltk==3.9.1
openai==0.28.1
propcache==0.3.1
regex==2024.11.6
requests==2.32.3
textblob==0.19.0
tqdm==4.67.1
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
